requirements.txt
atoma~=0.0.17
requests~=2.28.0
python-devlog~=1.2
qbittorrent-api~=2022.11.40

num2words~=0.5.12
aniparse~=1.1.2
Anisearch~=1.1.0

pymediainfo~=5.1.0
